feat: add private and public keys on duffle init
This change causes a fresh duffle init to create keyrings for public and private keys. If no private key is provided, this creates a new private key and adds it to the private ring. If existing keyrings exist, they are left alone.

To test:
- Run `helm init`
- Verify the files in `$HOME/.duffle/*.ring`

You can use `gpg --list-packets $FILE` to verify that the files are in the standard PGP format.