feat: add private and public keys on duffle init (#264)
* feat: add private and public keys on duffle init

This change causes a fresh duffle init to create keyrings for public and private keys. If no private key is provided, this creates a new private key and adds it to the private ring. If existing keyrings exist, they are left alone.

To test:
- Run `helm init`
- Verify the files in `$HOME/.duffle/*.ring`

You can use `gpg --list-packets $FILE` to verify that the files are in the standard PGP format.

* feat: add public keys too

* fix: set mode on secret keyring to 0600

* fix: don't create secret ring until after passwords are verified.

* feat: add 'duffle key' commands

* fix: validate and load bundle in 'duffle key sign'

* fix: Fixed duplicate import println on anonymous keys