Add injected-files into invocation images
This allows invocation images requiring settings files to work
correctly.
e-g: for docker-app invocation image, you can use settings files to
provide templated parameters. This is important, as this invocation
image as a setting management that is case-sensitive, and depending on
which docker-app package you want to deploy, the settings to pass can be
different.
This would also make it easier to create a generic Helm invocation image
(e.g.: we can mount settings files, or even an helm chart in tarball
format to the invocation image)

Signed-off-by: Simon Ferquel <simon.ferquel@docker.com>