[Docker Driver] Fix mounts on Docker for Windows (#297)
There are actually 3 fixes in this one:
- first temporary directory on Windows failed (/tmp makes no sense on
Windows). If goos is Windows, we pass "" as temp dir to let the OS
choose which directory to choose.
- paths in which to mount the files in the containers could previously
be relative, which makes no sense (relative paths resolution from the client makes no sense from the container root FS)
- paths manipulation for putting things inside the container where done
using filepath package. This means that on a Windows client, we generate
paths in a Windows fashion (c:\var\data) even if we know that invocation
images are linux containers (assumption given by the fact we mount
/var/run/docker.sock which makes no sense on Windows containers)

Signed-off-by: Simon Ferquel <simon.ferquel@docker.com>