feat: sign bundles during 'duffle build'
Instead of 'duffle build' creating a cnab/bundle.json, this causes build to create a 'cnab/bundle.cnab', which is a signed JSON document. By default, it is signed with the first secret key in the ring. But it can be signed with any singing key in the ring by supplying the -u flag.