feat: add location to bundle parameters
This change adds the concept of a location to parameters in a bundle.json file. Parameters now act like credentials, where destinations can be specified either as environment variables or as paths.

This builds on @simonferquel 's work in a previous PR (#298)

BREAKING: This changes the name of one flag, removes another flag, adds a new section to bundle.json parameters, and adds 'omitempty' on a few struct values in bundle.json

Closes #294