feat: add 'duffle run' command
This adds a new 'duffle run' command, as well as a description (in bundle.json) of what additional actions are available.

To test:

- Modify bundle.json to include an 'actions' section
- Run 'duffle run ACTION RELEASE'

Runs are allowed to have parameters and credentials. They can only be run on bundles that have already been installed. If they modify the release, they MUST set 'modifes: true' in the bundle.json definition