feat: add interactive prompts to credential set generator
This is an attempt to make `duffle credentials generate` easier to use. This prompts the user to enter values for each credential as they go, instead of requiring the user to merely edit the credential file by hand.

To test, run `duffle creds gen` on your favorite bundle. Use the new `--dry-run` flag to test the prompt without writing the file.