Custom dockerfiles and build contexts
This adds the possibility to specify a custom build context and
Dockerfile for docker components. This makes it easier for adding duffle
compatibility in existing projects.
E.G.: We can now add a simple duffle.json file in the existing
docker/app code base that builds the invocation image from the same
Build context as before, but with a different Dockerfile

Signed-off-by: Simon Ferquel <simon.ferquel@docker.com>