ref(cmd/duffle): save signed bundle to local store
Currently, if an output-file is specified, a signed bundle
file will be saved to the local filesystem at the specified
path but not to the local store. This change ensures that
the bundle is saved to local store regardless of output-file.
A flag should not change the main outcome of the command; it
should enhance functionality.