Add Push/Pull from an image registry
Introduced `duffle bundle push/pull` to work with a registry,
leveraging genuinetools/reg.

For now we create a Manifest with a config mediatype fixed to docker
image, because registries only accept a limited number of config types
in their validation logic. Once they accept custom config types, we will
be able to change that.

Also, cross-repository references cannot be modelled in a registry
manifest, which makes it impossible to create images references at the
manifest level. We need to dig into that with docker/distribution team.

Signed-off-by: Simon Ferquel <simon.ferquel@docker.com>