fix(packager): import to dir in dest not temp dir
Before, we were unpacking a bundle to a temp directory,
then loading it to validate the bundle. This caused
some challenges on windows around file permissions.

This PR changes the import logic so that a bundle is
unpacked in a directory at the destination specified
called <bundle-name>-<version>/ instead and then
validated. that directory is removed if the bundle
is not valid. If a directory already exists with that
name, it is overwritten.

resolves #458