fix(*): fix pull image on export issue
We were not waiting for the image to be pulled
before going on to the next step and exiting the
archiveImage function. This caused the image to
not be pulled. There used to be a sleep as a temp
fix to the problem but sleep is fragile and never
was intended as a long term solution.

This solution logs the image pull output ensuring
that we wait until the image is pulled before
moving on to the next step which is saveing the
image archive.

It also adds a verbose flag to the export cmd which
when enabled shows users the path of the logs file
for that export.

resolves #468