fix(*): properly load images on import
Import was not properly loading images because
the filepath.Walk function was opening the
archives directory like it was a file and then
passing that into the ImageLoad function which
failed silently.

This code change does two things:
1. It checks os.FileInfo to see if the path is
a directory and returns if it is. If the path is not
a directory, it proceeds to import the image archive
file.
2. It adds a verbose flag to the import command. If
enabled, it streams the ouput from image load to stdout.

resolves #469