Proposal: Replace env var and file inputs with JSON on STDIN
This is an implementation of a proposal to change the configuration
input process for invocation images. In the existing spec, installation
details, image maps, parameters and credentials are mapped on to
environment variables or files that are then injected into the
invocation image when an action is executed.

This proposal seeks to remove that input mapping process and replace it
with a singular input procedure (passing JSON into STDIN). By proposing
such a change, we lose the immediate mapping to environment variables or
files that might have benefitted some bundle authors, but we gain a
simplified and unified configuration input process.

As a result of this change, we should be able to remove those portions
of the spec that are purely about mapping inputs to environment
variables or files. This also opens up the possibility to extend the
existing parameters API to fully support all of the types in JSONSchema
that could not cleanly serialize to an environment variable or file.