remove un-used search command
search command was implemented in the first iteration
of deadling with bundles and bundle repos. The search
command was left in a hidden state so that we could reuse
it later if necessary but the strategy for storing bundles
has change significantly. We may also consider moving the
search command under the bundle command group in the future
if/when we re-implement. For now, we'll remove it so we can
make the duffle code understandable without knowing historical
context.