Use standard Go 1.12 image for docker builds
This allows Go 1.12 features to be used.

An alternative would be to bump quay.io/deis/lightweight-docker-go to use a
later version of Go, but it's probably safer to get out of that game and use
the more widely supported golang image instead. Also it's not obvious how to
make quay.io/deis/lightweight-docker-go usable by all maintainers.

Note that SKIP_DOCKER=true is currently undocumented and so there is no action
for users to take. (Those who use SKIP_DOCKER=true should know what they are
doing.)