Rename k8s driver parm from KUBE_CONFIG to KUBECONFIG
Implement this by bumping cnab-go. Also picks up other fixes.

KUBECONFIG is the de facto standard environment variable for kubectl etc.

Without this change when the user omits KUBE_CONFIG, the default kube config
will be used, which may well point to a valid cluster, though this will
effectively be the wrong cluster when the user has set $KUBECONFIG.

This is a particular issue when kind (https://kind.sigs.k8s.io/) is used as
it doesn't update the default kube config.